+++ packages/math-input/src/components/keypad-v1/__tests__/node-manager.test.ts
@@ -1,4 +1,4 @@
-import NodeManager from "../keypad-v1/node-manager";
+import NodeManager from "../node-manager";
 
 describe("NodeManager", () => {
     let nodeManager;
